// src/services/Reline/interfaces/graph.relineSchema.ts
import { RelineElementBaseSchema } from "./elementBase.relineSchema";
import { RelineNodeSchema } from "./node.relineSchema";
import { RelineEdgeSchema } from "./edge.relineSchema";
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface RelineGraphSchema extends RelineElementBaseSchema{
  id: string;
  metadata: {
    basic: {
      id: string;
      RelineGraphName?: string;
      description?: string;
      keywords?: string[];
      repositories?: {
        type?: string;
        url?: string;
        [k: string]: unknown;
      };
      license?: string;
      version?: string;
      contributors?: {
        name?: string;
        email?: string;
        description?: string;
        [k: string]: unknown;
      }[];
      [k: string]: unknown;
    };
    [k: string]: {
      [k: string]: {
        [k: string]: any
      };
    }|any;
  };
  nodes: {
    [k: string]: RelineNodeSchema;
  };
  edges: RelineEdgeSchema[];
}